Fix 32bit compilation

+++ src/main.cpp
@@ -11,7 +11,7 @@ void parse_oui_data()
 	std::string str;
 
 	t.seekg(0, std::ios::end);
-	str.reserve(t.tellg());
+	str.reserve(size_t(t.tellg()));
 	t.seekg(0, std::ios::beg);
 
 	str.assign((std::istreambuf_iterator<char>(t)), std::istreambuf_iterator<char>());
